небольшая оптимизация по размерам

+++ src/AspNetCoreTest.Data.Models/WSRequest.cs
@@ -47,8 +47,8 @@ namespace AspNetCoreTest.Data.Models
         public int LenZ { get; set; }
 
         // статистика по весам (если есть конечно) (для отрисовки тока)
-        public double MinWeight { get; set; }
-        public double MaxWeight { get; set; }
+        public float MinWeight { get; set; }
+        public float MaxWeight { get; set; }
 
         // пороговые значения нейронов (для отрисовки тока)
         public int MaxState { get; set; }
